requirements.txt
Django==1.9.2
pinax-theme-bootstrap==7.2.0
django-user-accounts==1.3.1
metron==1.3.7
pinax-eventlog==1.1.1
django-toolbelt
dj-database-url
django-localflavor
pyTelegramBotAPI
